Lift fake output and pdf
